Move 'SQLTerm` to generic `CodeFormatter`
