+++ src/main/scala/temple/generate/utils/CodeFormatter.scala
@@ -1,11 +1,11 @@
-package temple.generate.database
+package temple.generate.utils
 
 import temple.utils.StringUtils.indent
 
 import scala.Option.when
 
 /** Any nested sequence of strings that can be flattened and iterated */
-trait SQLTerm {
+trait CodeFormatter {
 
   /**
     * Iterate through any ordered structure of strings
@@ -14,54 +14,55 @@ trait SQLTerm {
   def flatIterator: Iterator[String]
 }
 
-object SQLTerm {
+object CodeFormatter {
 
-  /** Turns a [[String]] into a [[temple.generate.database.SQLTerm]], with a trivial iterator of just the single string
+  /** Turns a [[String]] into a [[CodeFormatter]], with a trivial iterator of just the single string
     */
-  implicit class SQLTermString(string: String) extends SQLTerm {
+  implicit class CodeFormatterString(string: String) extends CodeFormatter {
     override def flatIterator: Iterator[String] = Iterable.single(string).iterator
   }
 
   /**
     * Turns a list of items into a string with separators, constructed with
-    * [[temple.generate.database.SQLTerm.mkSQL#list(temple.generate.database.SQLTerm.SQLTermIterableString)]]
+    * [[utils.CodeFormatter.mkSQL#list(temple.generate.database.CodeFormatter.CodeFormatterIterableString)]]
+    *
     * @param strings a list of strings to be separated
     * @param separator the separator between the items
     */
-  class SQLTermList(strings: IterableOnce[String], separator: String = ",") extends SQLTerm {
+  class CodeFormatterList(strings: IterableOnce[String], separator: String = ",") extends CodeFormatter {
 
     override def flatIterator: Iterator[String] = strings.iterator.zipWithIndex.map {
       case (term, i) => mkSQL(when(i != 0)(separator), term)
     }
 
     /** Add newlines after each separator */
-    def spaced: SQLTermList = new SQLTermList(strings, ",\n")
+    def spaced: CodeFormatterList = new CodeFormatterList(strings, ",\n")
   }
 
-  /** Turns a [[IterableOnce]] of [[SQLTerm]]s into a [[temple.generate.database.SQLTerm]], with an iterator that visits
+  /** Turns a [[IterableOnce]] of [[CodeFormatter]]s into a [[CodeFormatter]], with an iterator that visits
     * every child */
-  implicit class SQLTermIterable(strings: IterableOnce[SQLTerm]) extends SQLTerm {
+  implicit class CodeFormatterIterable(strings: IterableOnce[CodeFormatter]) extends CodeFormatter {
     override def flatIterator: Iterator[String] = strings.iterator.flatMap(_.flatIterator)
   }
 
   /** An iterable of strings can be treated as an iterable of SQLTerms, but can also be turned into a list. */
-  implicit class SQLTermIterableString(strings: IterableOnce[String])
-      extends SQLTermIterable(strings.iterator.map(SQLTermString)) {
+  implicit class CodeFormatterIterableString(strings: IterableOnce[String])
+      extends CodeFormatterIterable(strings.iterator.map(SQLTermString)) {
 
     /** Turn a list of SQL terms into a comma-separated list */
-    def mkSQLList: SQLTermList = new SQLTermList(strings)
+    def mkSQLList: CodeFormatterList = new CodeFormatterList(strings)
   }
 
   object mkSQL {
 
     /** Turn a list of terms into a comma-separated string */
-    def list(terms: SQLTermIterableString): String = mkSQL(terms.mkSQLList)
+    def list(terms: CodeFormatterIterableString): String = mkSQL(terms.mkSQLList)
 
     /** Turn a list of terms into a comma- and newline-separated string */
-    def spacedList(terms: SQLTermIterableString): String = mkSQL(terms.mkSQLList.spaced)
+    def spacedList(terms: CodeFormatterIterableString): String = mkSQL(terms.mkSQLList.spaced)
 
     /** Combine a sequence of SQL terms into a single string, omitting spaces as necessary */
-    def apply(strings: SQLTerm*): String = {
+    def apply(strings: CodeFormatter*): String = {
       val iterator: Iterator[String] = strings.iterator.flatMap(_.flatIterator)
 
       // Iterate through the strings with a string builder, keeping track of whether the previous segment ends in a
@@ -82,16 +83,16 @@ object SQLTerm {
     }
 
     /** Construct a SQL statement, like in
-      * [[temple.generate.database.SQLTerm.mkSQL#apply(scala.collection.immutable.Seq)]] but ending in a semicolon */
-    def stmt(string: SQLTerm*): String = mkSQL(string, ";")
+      * [[utils.SQLTerm.mkSQL#apply(scala.collection.immutable.Seq)]] but ending in a semicolon */
+    def stmt(string: CodeFormatter*): String = mkSQL(string, ";")
   }
 
   object sqlParens {
 
     /** Wrap a SQL snippet in parentheses */
-    def apply(string: SQLTerm*): String = mkSQL("(", string, ")")
+    def apply(string: CodeFormatter*): String = mkSQL("(", string, ")")
 
     /** Wrap a SQL snippet in parentheses, with newlines inside them */
-    def spaced(string: SQLTerm*): String = mkSQL("(", "\n", indent(mkSQL(string)), "\n", ")")
+    def spaced(string: CodeFormatter*): String = mkSQL("(", "\n", indent(mkSQL(string)), "\n", ")")
   }
 }
